refactoring scraper
