+++ lib/manager_save.rb
@@ -1,0 +1,8 @@
+require 'activity_logger'
+class ManagerSave 	
+ 	def save(site_name, my_hash)
+ 		conn = ActivityLogger.new
+		method_name = "save_scrap_#{site_name}"
+		conn.send(method_name,my_hash)
+	end
+end
